# backend/docs/swagger.json
{
  "definitions": {
...
    "schemas.JWT": {
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "schemas.Response": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "type": "object"
    }
  },
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/auth/token": {
      "post": {
        "description": "Authenticates a user and provides a JWT to Authorize API calls",
        "operationId": "Authentication",
        "parameters": [
          {
            "description": "Username",
            "in": "formData",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "description": "Password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/schemas.JWT"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/schemas.Response"
            }
          }
        },
        "summary": "Provides a JSON Web Token"
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}